% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package_versions.R
\name{check_package_versions}
\alias{check_package_versions}
\title{Check package versions between server collection and those installed}
\usage{
check_package_versions()
}
\description{
Where applicable gives the option to upgrade installed packages from server
and to downgrade and install the server version where currently exceeded.
}
\examples{
\dontrun{
check_package_versions()
}
}
