% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_server_path.R
\name{get_server_path}
\alias{get_server_path}
\title{Get the current path used to find R packages.}
\usage{
get_server_path()
}
\value{
Char of server path.
}
\description{
This should be the directory containing Windows zipped binary files.
}
\examples{
\dontrun{
get_server_path()
}
}
