% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installer.R
\name{package_installer}
\alias{package_installer}
\title{Install R package from internal server}
\usage{
package_installer()
}
\description{
To be used in R Studio, optionally as an add-in. A pop-up window will ask
to enter name of package wish to install. If wish to enter more than one
package separate the names with commas
}
\examples{
\dontrun{
package_installer()
}
}
