% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_server_path.R
\name{update_server_path}
\alias{update_server_path}
\title{Set server path containing packages to install.}
\usage{
update_server_path()
}
\description{
This is only needed if get_server_path() shows current path needs changing.
It will only need to be set once as the path specified will be stored (
inside the package install) for all subsequent uses of the
package_installer() function.
}
\examples{
\dontrun{
update_server_path()
}
}
